COMP: removing conflicting return type with parent class

+++ BlueBerry/Testing/org.blueberry.ui.tests/src/api/berryMockSelectionProvider.cpp
@@ -42,7 +42,7 @@ struct EmptyStructuredSelection : public IStructuredSelection
     return elements->end();
   }
 
-  size_t Size() const
+  int Size() const
   {
     return elements->size();
   }
